📑 :: comment utils 수정

+++ project/utils/comment.py
@@ -3,9 +3,10 @@ from sqlalchemy.orm import Session
 from project.utils.security import check_comment
 
 
-def make_comment(user_id: int, content: str, session: Session):
+def write_comment(user_id: int, content: str,feed_id: int, session: Session):
     new_comment = Comment(
         content=content,
+        feed_id=feed_id,
         user_id=user_id
     )
 
@@ -18,7 +19,8 @@ def make_comment(user_id: int, content: str, session: Session):
 
 def delete_comment(user_id: int, comment_id: int, session: Session):
     comment = check_comment(comment_id=comment_id, user_id=user_id, session=session)
-    session.delete(comment)
+    if comment:
+        session.delete(comment)
 
     return {
         "message": "success"
@@ -26,9 +28,8 @@ def delete_comment(user_id: int, comment_id: int, session: Session):
 
 def modify_comment(user_id: int, content: str, comment_id: int , session: Session):
     comment = check_comment(comment_id=comment_id, user_id=user_id, session=session)
-    if comment == True:
-        modify_comment = session.query(content).filter()
-        modify_comment.update_one({content: content})
+    if comment:
+        comment.content = content
 
         return {
             "message": "success"
